Added GLUT_CURSOR_NONE support in UNIX_X11 (well, NetBSD; you lot need to try it on others; (^&).
Deallocated some resources that we are creating.  VERY slight memory leak,
but plugged now.


These two complete the first two "Open issues" ( (a) and (b) ).  The first
one also completes X support for glutSetCursor().  If others can verify,
we can fully demote the outstanding bug over this to WIN32-specific.

I'll delete the dead Open issues and re-letter the others if someone will
cross-check me.  Or if no one says anything in a day or two.  (^&


git-svn-id: https://freeglut.svn.sourceforge.net/svnroot/freeglut/trunk/freeglut@310 7f0cb862-5218-0410-a997-914c9d46530a