Fixed a bug for UNIX_X11 where window reshape events were causing freeglut to mark the window as *needing* a reshape (which during general display callback handling would result in effectively a glutReshapeWindow()).
The code is now system-dependant.  It should be abstracted to a function,
but is presently copied in two places.  Sorry.

Also, inverted the order of the associated if()/else check (in both
WIN32 and UNIX_X11 branches) since the former "else" part was a one-line
callback invocation.

This fixes two seemingly unrelated bugs that I was seeing in UNIX_X11.


git-svn-id: https://freeglut.svn.sourceforge.net/svnroot/freeglut/trunk@374 7f0cb862-5218-0410-a997-914c9d46530a